new class Tool for equipable item :hammer:

+++ ShakeAndCraft/app/src/main/java/com/example/shakecraft/model/Boss.kt
@@ -3,14 +3,14 @@ package com.example.shakecraft.model
 
 class Boss (
     var name: String,
-    var life: Int,
+    var life: Double,
     var maxlife: Int,
     var image: Int,
     var xpReward: Int,
     val possibleLoot: List<Pair<Item, Double>>,
 ){
 
-    fun takeDamage(strength: Int) {
+    fun takeDamage(strength: Double) {
         this.life -= strength
     }
 }
